% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_wrangling.R
\name{do_coord_conversion}
\alias{do_coord_conversion}
\title{do_coord_conversion}
\usage{
do_coord_conversion(northings, eastings)
}
\arguments{
\item{northings}{Numeric vector of northings to be converted.}

\item{eastings}{Numeric vector of eastings to be converted.}
}
\value{
A list object containing two vectors, one for latitude and one for longitude.
}
\description{
Helper function to convert northings/eastings to latitude/longitude respectively.
}
\examples{
\dontrun{
do_coord_conversion(
 292500,
 72500
)
}

}
