% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{download_ukcp_csv_results}
\alias{download_ukcp_csv_results}
\title{download_ukcp_csv_results}
\usage{
download_ukcp_csv_results(
  status_url,
  output_folder = "outputs",
  keep_zip = F,
  API_KEY = Sys.getenv("UKCP_API_KEY")
)
}
\arguments{
\item{status_url}{URL used to retrieve results from the API.}

\item{output_folder}{Folder (in working directory) to save .csv files to.}

\item{keep_zip}{Keep the original zip file (in your working directory root folder). Note the initial name will always be \code{temp.zip}.}

\item{API_KEY}{UKCP API key. Defaults to using environment variable \code{UKCP_API_KEY}.}
}
\value{
Downloads the .zip file held at the \code{status_url} and extracts the csv(s) within (users can optionally keep the .zip file).
}
\description{
Download the results of your API query, using the status_url returned in \code{start_ukcp_csv_job}. Polls server once every five seconds.
}
\examples{
\dontrun{
 download_ukcp_csv_results(
  status_url = "https://ukclimateprojections-ui.metoffice.gov.uk/status/your-id-here",
  output_folder = "outputs"
)
}

}
