% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_wrangling.R
\name{read_projection_csv}
\alias{read_projection_csv}
\title{read_projection_csv}
\usage{
read_projection_csv(filepath, convert_coordinates = T)
}
\arguments{
\item{filepath}{Location of the file.}

\item{convert_coordinates}{Convert coordinates from northings/eastings to latitude/longitude.}
}
\value{
A dataframe with the processed data (and pertinent metadata) in tall format.
}
\description{
Read a UKCP projection data .csv file(s) and handle some basic processing.
}
\examples{
\dontrun{
read_projection_csv(
  "filename.csv"
)
}

}
