% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{start_ukcp_observation_csv_job}
\alias{start_ukcp_observation_csv_job}
\title{start_ukcp_observation_csv_job}
\usage{
start_ukcp_observation_csv_job(
  label,
  API_KEY = Sys.getenv("UKCP_API_KEY"),
  identifier,
  area_type,
  area_details,
  collection,
  temp_avg,
  year_range,
  variable
)
}
\arguments{
\item{label}{Job label for your request - this will appear in the UKCP API.}

\item{API_KEY}{Your UKCP API key. Defaults to the \code{UKCP_API_KEY} environment variable if created.}

\item{identifier}{Identifier for the dataset you would like to query.}

\item{area_type}{How area you would like to cover is set up, eg a 'point', 'admin_region' or a 'bbox' (these are not the full set of options) to cover multiple grid cells.}

\item{area_details}{The name or coordinates (northings/eastings) of the location you are interested in. If coordinates, the the point ID or edges of your boundary. In the latter case, enter as a vector with W, S, E, N boundaries in that order.}

\item{collection}{The collection the queried dataset is part of.}

\item{temp_avg}{The time period the data is broken down by, eg 'day' or 'may' (for data for that month, 'mall' for all months).}

\item{year_range}{Years covered by your call. Enter as a vector in the following format: \code{c(min_yr, max_yr)}.}

\item{variable}{The measurement you are trying to extract, eg 'rainfall' or 'sun'.}
}
\value{
A list of information about your job, crucially the URL to use to get the results (when ready).
}
\description{
Send a query to the UKCP API for observation data and get a job (that will produce .csv outputs) started.
}
\examples{
\dontrun{
 start_ukcp_observation_csv_job(
  label = "ukcpR-query",
  identifier = "LS6_Subset_02",
  area_type = "point",
  area_details = c(292500, 72500),
  collection = "land-obs_5",
  temp_avg = "day",
  year_range = c(1960, 2021),
  variable = "rainfall"
 )
 }

}
