% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{start_ukcp_projection_csv_job}
\alias{start_ukcp_projection_csv_job}
\title{start_ukcp_projection_csv_job}
\usage{
start_ukcp_projection_csv_job(
  label,
  API_KEY = Sys.getenv("UKCP_API_KEY"),
  identifier,
  area_details,
  collection,
  temp_avg,
  year_range,
  variable,
  climate_change_type,
  ensemble_member_set,
  scenario
)
}
\arguments{
\item{label}{Job label for your request - this will appear in the UKCP API.}

\item{API_KEY}{Your UKCP API key. Defaults to the \code{UKCP_API_KEY} environment variable if created.}

\item{identifier}{Identifier for the dataset you would like to query.}

\item{area_details}{The coordinates (northings/eastings) of the location you are interested in. Enter as a vector with W, S, E, N boundaries in that order.}

\item{collection}{The collection the queried dataset is part of.}

\item{temp_avg}{The time period the data is broken down by, eg 'day' or 'may' (for data for that month, 'mall' for all months).}

\item{year_range}{Years covered by your call. Enter as a vector in the following format: \code{c(min_yr, max_yr)}. Note there may be restrictions on which years can be sent depending on which data you are requesting.}

\item{variable}{The measurement you are trying to extract, eg 'rainfall' or 'sun'.}

\item{climate_change_type}{Whether you want 'absolute' or 'anomaly' values.}

\item{ensemble_member_set}{Set of models used to generate projections.}

\item{scenario}{RCP scenario (rcp26 or rcp85) to gather projections for. Note only variable for global projections.}
}
\value{
A list of information about your job, crucially the URL to use to get the results (when ready).
}
\description{
Send a query to the UKCP API for projection data and get a job (that will produce .csv outputs) started.
}
\examples{
\dontrun{
 start_ukcp_projection_csv_job(
  label = "ukcpR-query",
  identifier = "LS2_Subset_01",
  area_details = c(292500, 72500, 293500, 73500),
  collection = "land-gcm",
  temp_avg = "mall",
  year_range = c(2061, 2080),
  variable = "tas"
 )
 }

}
