% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcpR-package.R
\docType{package}
\name{ukcpR-package}
\alias{ukcpR}
\alias{ukcpR-package}
\title{ukcpR: Work with the Met Office UKCP API via R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package lets you interact with the UK Met Office UKCP API using R. It is primarily aimed at extracting and manipulating .csv exports (at least for now).
}
\keyword{internal}
